Fix Affordable Housing edge case
